File restructuring and unit test fixes
