+++ dpypelines/pipeline/functions/schemas.py
@@ -8,7 +8,7 @@ def get_config_schema_path(config: dict) -> Path:
     Returns the local path to a schema from the `$id` specified in a pipeline config dictionary
     """
     # Set the base path for the schemas
-    schema_base_path = Path("schemas/dataset-ingress/config").absolute()
+    schema_base_path = Path("dpypelines/schemas/dataset-ingress/config").absolute()
 
     # Check `$id` is in the config dictionary keys
     if "$id" not in config.keys():
