+++ dpypelines/pipeline/shared/details.py
